Mise en place de la scène de fin
Présentation des stats de la partie.

Pour tester sans avoir à attendre tout le temps de la musique, il faut
décommenter la ligne dans le MusicManager afin de retourner une durée de
musique de 15 secondes

+++ Unity App/Assets/Scripts/GUIManager.cs
@@ -68,7 +68,7 @@ public class GUIManager : MonoBehaviour
 		instance.multiplicateurlabel.text = "Multiplicateur: "+multi;
 	}
 	
-	public static void SetPourcentage(int pourcents)
+	public static void SetPourcentage(float pourcents)
 	{
 		instance.pourcentslabel.text = pourcents+"%";
 	}
